requirements.txt
openai == 1.6.1
python-dotenv == 1.0.0